change count variable to unsigned long for large valum input

+++ include/caffe/layers/split_layer.hpp
@@ -37,7 +37,7 @@ class SplitLayer : public Layer<Dtype> {
   virtual void Backward_gpu(const vector<Blob<Dtype>*>& top,
       const vector<bool>& propagate_down, const vector<Blob<Dtype>*>& bottom);
 
-  int count_;
+  unsigned long count_;
 };
 
 }  // namespace caffe
